Move the non unit tests to `tool_use`
